app/requirements.txt
asgiref==3.9.2
cloudpickle==3.1.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
joblib==1.5.2
llvmlite==0.45.1
npm==0.1.1
numba==0.62.1
numpy==2.3.3
nvidia-nccl-cu12==2.28.3
openpyxl==3.1.5
optional-django==0.1.0
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
shap==0.48.0
six==1.17.0
slicer==0.0.8
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
xgboost==3.0.5
